requirements.txt
streamlit
yfinance
ta
python-telegram-bot
streamlit-autorefresh